requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
hl7==0.4.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.16.4
pytest==8.4.1
httpx==0.28.1
